Added all Play versions up to 2.8
