--- play24/play-monad/src/main/scala/dev/playmonad/package.scala
@@ -1,17 +1,0 @@
-package dev
-
-import controllers.Assets.BadRequest
-import play.api.mvc.BodyParser
-
-import scala.concurrent.Future
-
-package object playmonad {
-  def header(name: String): HeaderReader.Aux[String] = HeaderReader { request =>
-    request.headers.get(name) match {
-      case Some(value) => Future.successful(Right(value))
-      case None        => Future.successful(Left(BadRequest(s"Header $name must be provided")))
-    }
-  }
-
-  def body[A](bodyParser: BodyParser[A]): BodyReader.Aux[A, A] = BodyReader(bodyParser)
-}
