+++ project/ScalacOptions.scala
@@ -158,7 +158,7 @@ object ScalacOptions {
     "-language:implicitConversions", // Allow definition of implicit functions called views
     "-unchecked", // Enable additional warnings where generated code depends on assumptions.
     "-Xcheckinit",      // Wrap field accessors to throw an exception on uninitialized access.
-    "-Xfatal-warnings", // Fail the compilation if there are any warnings.
+//    "-Xfatal-warnings", // Fail the compilation if there are any warnings.
 //    "-Xfuture",            // Turn on future language features.
     "-Xlint:adapted-args", // Warn if an argument list is modified to match the receiver.
 //    "-Xlint:by-name-right-associative", // By-name parameter of right associative operator.
