+++ sample-play25/project/build.properties
