+++ sample-play25/project/plugins.sbt
